+++ Runtime/Extensions/Color/ColorAdjustmentExtensions.cs
@@ -1,21 +1,10 @@
 using LiteNinja.Colors.Spaces;
-using LiteNinja.Utils.Extensions;
 using UnityEngine;
 
-namespace LiteNinja.Colors.extensions
+namespace LiteNinja.Colors.Extensions
 {
-    public static class ColorExtensions
+    public static class ColorAdjustmentExtensions
     {
-        public static readonly Color transparentBlack = new Color(0f, 0f, 0f, 0f);
-
-        public static float Brightness(this Color self)
-        {
-            const float oneOver3 = 1f / 3f;
-            return (self.r + self.g + self.b)*oneOver3;
-        }
-
-        #region Create a color changing one parameter at a time
-
         public static Color WithAlpha(this Color self, float alpha)
         {
             return new Color(self.r, self.g, self.b, alpha);
@@ -68,13 +57,22 @@ namespace LiteNinja.Colors.extensions
             hsv.Value = value;
             return hsv;
         }
+        
+        public static Color WithRelativeLuminance(this Color self, float relativeLuminance)
+        {
+            //find color channels for relative luminance
+            var r = (float)(relativeLuminance <= 0.03928 ? 12.92 * relativeLuminance : Mathf.Pow((relativeLuminance + 0.055f)/1.055f, 1.0f/2.4f));
+            var g = (float)(relativeLuminance <= 0.03928 ? 12.92 * relativeLuminance : Mathf.Pow((relativeLuminance + 0.055f)/1.055f, 1.0f/2.4f));
+            var b = (float)(relativeLuminance <= 0.03928 ? 12.92 * relativeLuminance : Mathf.Pow((relativeLuminance + 0.055f)/1.055f, 1.0f/2.4f));
+            
+            //return color with relative luminance
+            return new Color(r, g, b, self.a);
+        }
 
         public static Color Opaque(this Color color)
         {
             return new Color(color.r, color.g, color.b);
         }
-
-        #endregion
 
         public static Color Complementary(this Color self)
         {
@@ -87,36 +85,8 @@ namespace LiteNinja.Colors.extensions
         {
             return new Color(1f - self.r, 1f - self.g, 1f - self.b, self.a);
         }
-
-        #region Create colors by changing parameters
-
-        public static Color AddBrightness(this Color self, float amount)
-        {
-            return new Color(Mathf.Clamp01(self.r + amount), Mathf.Clamp01(self.g + amount),
-                Mathf.Clamp01(self.b + amount), self.a);
-        }
-
-        public static Color AddContrast(this Color self, float amount)
-        {
-            return new Color(Mathf.Clamp01((1f + amount) * (self.r - 0.5f) + 0.5f),
-                Mathf.Clamp01((1f + amount) * (self.g - 0.5f) + 0.5f),
-                Mathf.Clamp01((1f + amount) * (self.b - 0.5f) + 0.5f),
-                self.a);
-        }
-
-        public static Color AddSaturation(this Color self, float amount)
-        {
-            ColorHSV colorHsv = self;
-            colorHsv.Saturation += amount;
-            return colorHsv;
-        }
-
-        public static Color AddHue(this Color self, float amount)
-        {
-            ColorHSV colorHsv = self;
-            colorHsv.Hue += amount;
-            return colorHsv;
-        }
+        
+        
 
         public static Color AddLightness(this Color self, float amount)
         {
@@ -169,38 +139,7 @@ namespace LiteNinja.Colors.extensions
                 Mathf.Clamp01(self.b + yellowBlue - (cyanRed + magentaGreen) * 0.5f),
                 self.a);
         }
-
-        #endregion
-
-
-        #region Color comparison
-
-        public static bool Approximately(this Color value, Color other)
-        {
-            return Mathf.Approximately(value.r, other.r) &&
-                   Mathf.Approximately(value.g, other.g) &&
-                   Mathf.Approximately(value.b, other.b) &&
-                   Mathf.Approximately(value.a, other.a);
-        }
-
-        public static bool Approximately(this Color value, Color other, float epsilon)
-        {
-            return value.r.Approximately(other.r, epsilon) &&
-                   value.g.Approximately(other.g, epsilon) &&
-                   value.b.Approximately(other.b, epsilon) &&
-                   value.a.Approximately(other.a, epsilon);
-        }
-
-        public static bool IsApproximatelyBlack(this Color self)
-        {
-            return self.r + self.g + self.b <= Mathf.Epsilon;
-        }
-
-        public static bool IsApproximatelyWhite(this Color self)
-        {
-            return self.r + self.g + self.b >= 1 - Mathf.Epsilon;
-        }
-
-        #endregion
+        
+        
     }
 }