+++ Runtime/Helpers/ColorVisionType.cs
@@ -1,0 +1,15 @@
+namespace LiteNinja.Colors.Helpers
+{
+    public enum ColorVisionType
+    {
+        Normal,
+        Protanopia,
+        Protanomaly,
+        Deuteranopia,
+        Deuteranomaly,
+        Tritanopia,
+        Tritanomaly,
+        Achromatopsia,
+        Achromatomaly
+    }
+}