+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Color Extensions for Contrast and Hue
+- Color Extensions for Saturation and Lightness
+- Color Extensions for Color blindness
+
+### Changed
+- Split ColorExtensions in multiple files by common theme
 
 ## [1.1.2] - 2022-07-07
 ### Added